requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
django-multiselectfield==0.1.13
djangorestframework==3.15.2
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
